+++ rocks/__init__.py
@@ -1,7 +1,7 @@
 """For space rocks."""
 
 # Welcome to rocks
-__version__ = "1.6.4"
+__version__ = "1.6.5"
 
 import inspect