+++ CHANGELOG.md
@@ -1,9 +1,10 @@
-# 1.6.5 - 2022-11-
+# 1.6.5 - 2022-11-08
 - Code refactored to optimize performance of CLI, e.g. $ rocks id is now 2x faster
 - identify() no longer accepts pd.Series to avoid heavy pandas import
 - Adding '--clear' and '--update' flags to $ rocks status
 - $ rocks author output now includes bibcode
 - Switching from warnings to logging module
+- Fix bug in datacloud families ingestion
 
 # 1.6.4 - 2022-10-21
 - Add 'author' command to quickly check the presence of data from peer-reviewed article: '$ rocks author bowell'